Move all calculations server side
