+++ src/components/data-tables/data-table-title.tsx
@@ -1,0 +1,5 @@
+﻿function DataTableTitle({ text }: { text: string }) {
+    return <h3 className="mb-3 text-lg font-semibold">{text}</h3>;
+}
+
+export default DataTableTitle;
